Reorganizado estrutura de pastas e arquivos.
